+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/http/msf4j/Msf4jHttpRequest.java
@@ -16,7 +16,7 @@
  * under the License.
  */
 
-package org.wso2.carbon.uis.internal.http;
+package org.wso2.carbon.uis.internal.http.msf4j;
 
 import io.netty.handler.codec.http.QueryStringDecoder;
 import io.netty.handler.codec.http.cookie.Cookie;
